[Feature] [Update] Test y Funciones de Backend
Se crearon los test y se modificaron ciertas funciones al punto de eliminarlas porque no cumplian ninguna función dentro del codigo y fuera del codigo (frontend)

+++ cliente/src/main/java/mingetal/MCM/cliente/Services/ClienteService.java
@@ -1,7 +1,7 @@
-package mingetal.MCM.cliente.Services;
+package mingetal.MCM.cliente.services;
 
-import mingetal.MCM.cliente.Entities.ClienteEntity;
-import mingetal.MCM.cliente.Repositories.ClienteRepository;
+import mingetal.MCM.cliente.entities.ClienteEntity;
+import mingetal.MCM.cliente.repositories.ClienteRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,26 +13,38 @@ public class ClienteService {
     @Autowired
     ClienteRepository clienteRepository;
 
-    public void save(ClienteEntity clienteEntity) {
-        String[] palabras = clienteEntity.getNombre().split("\\s+");
-        StringBuilder sb = new StringBuilder();
+    //-------------------- Guardado --------------------
 
-        for (String palabra : palabras) {
-            // Convertir la primera letra de la palabra a mayúscula y agregar el resto de la palabra
-            sb.append(Character.toUpperCase(palabra.charAt(0))).append(palabra.substring(1)).append(" ");
+    /*
+    *
+    *
+    *
+    * */
+    public ClienteEntity save(ClienteEntity clienteEntity) {
+        if(findByRut(clienteEntity.getRut())==null){
+            String[] palabras = clienteEntity.getNombre().split("\\s+");
+            StringBuilder sb = new StringBuilder();
+
+            for (String palabra : palabras) {
+                // Convertir la primera letra de la palabra a mayúscula y agregar el resto de la palabra
+                sb.append(Character.toUpperCase(palabra.charAt(0))).append(palabra.substring(1)).append(" ");
+            }
+            clienteEntity.setNombre(sb.toString().trim());
+            return clienteRepository.save(clienteEntity);
         }
-        clienteEntity.setNombre(sb.toString().trim());
-        clienteRepository.save(clienteEntity);
+        return null;
     }
+
+    //-------------------- Buscar --------------------
 
     public List<ClienteEntity> findAll() {
         List<ClienteEntity> clienteEntities = clienteRepository.findAll();
         clienteEntities.sort((p1, p2) -> p1.getNombre().compareTo(p2.getNombre()));
         return clienteEntities;
     }
-
-    public ClienteEntity findByRut(String rut) { return clienteRepository.findByRut(rut); }
-
+    public ClienteEntity findByRut(String rut) {
+        return clienteRepository.findByRut(rut);
+    }
     public List<ClienteEntity> findByNombre(String nombre) {
         List<ClienteEntity> clienteEntities = findAll();
         List<ClienteEntity> resultados = new ArrayList<>();
@@ -49,32 +61,29 @@ public class ClienteService {
     public ClienteEntity findByNombreTextual(String nombre) {
         return clienteRepository.findByNombre(nombre);
     }
-
     public List<ClienteEntity> findByEmpresa(String empresa) {
         List<ClienteEntity> clienteEntities = clienteRepository.findByEmpresa(empresa);
         clienteEntities.sort((p1, p2) -> p1.getNombre().compareTo(p2.getNombre()));
         return clienteEntities;
     }
 
+    //-------------------- Eliminar --------------------
+
     public ClienteEntity delete(String rut) {
         ClienteEntity clienteEntity = findByRut(rut);
+        if(clienteEntity==null){
+            return null;
+        }
         clienteRepository.delete(clienteEntity);
         return clienteEntity;
     }
 
-    public ClienteEntity deleteByNombre(String nombre) {
-        ClienteEntity clienteEntity = findByNombre(nombre).get(0);
-        clienteRepository.delete(clienteEntity);
-        return clienteEntity;
-    }
+    //-------------------- Editar --------------------
 
     public ClienteEntity update(ClienteEntity clienteEntity) {
-        ClienteEntity clienteEntity1 = findByRut(clienteEntity.getRut());
-        clienteEntity1.setNombre(clienteEntity.getNombre());
-        clienteEntity1.setEmail(clienteEntity.getEmail());
-        clienteEntity1.setTelefono(clienteEntity.getTelefono());
-        clienteEntity1.setEmpresa(clienteEntity.getEmpresa());
-        return clienteRepository.save(clienteEntity1);
+        if(findByRut(clienteEntity.getRut())==null){
+            return null;
+        }
+        return clienteRepository.save(clienteEntity);
     }
-
 }
